refactor: 集計ロジックを分離しJSフォールバックを廃止
- duckdbBridge.tsからSQL集計ロジックをduckdbAggregator.tsに抽出
- aggregate.tsからJS版集計関数を削除し型定義のみに
- main.tsからisWasmReady分岐とJSフォールバックパスを削除
